requirements.txt
# Core dependencies
rdkit-pypi>=2022.9.5
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.3
seaborn>=0.11.0
scikit-learn>=1.1.0

# Jupyter notebook support
jupyter>=1.0.0
ipykernel>=6.0.0

# Optional: for better performance
numba>=0.56.0

# Development dependencies (optional)
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0
